+++ healthcheck.go
@@ -3,21 +3,10 @@ package healthcheck
 import (
 	"fmt"
 	"net/http"
-
-	"git.aqq.me/go/app/appconf"
-	"git.aqq.me/go/app/applog"
-	"git.aqq.me/go/app/event"
-	"github.com/iph0/conf"
-	"go.uber.org/zap"
 )
 
-type healthcheckConfig struct {
-	Listen string
-}
-
-type server struct {
-	log      *zap.SugaredLogger
-	listener *http.Server
+type Handler struct {
+	http.ServeMux
 }
 
 // State type
@@ -39,88 +28,33 @@ var stateMap = map[State]int{
 	3: 500,
 }
 
-var srv *server
-
-func init() {
-	event.Init.AddHandler(
-		func() error {
-			cnf := appconf.GetConfig()["healthcheck"]
-
-			var config healthcheckConfig
-			err := conf.Decode(cnf, &config)
-			if err != nil {
-				return err
-			}
-
-			srv = &server{
-				log: applog.GetLogger().Sugar(),
-				listener: &http.Server{
-					Addr: config.Listen,
-				},
-			}
-
-			go func() {
-				err = srv.listener.ListenAndServe()
-				if err != nil && err != http.ErrServerClosed {
-					srv.log.Error(err)
-				}
-			}()
-
-			srv.log.Info("Started healthcheck listener")
-
-			return nil
-		},
-	)
-
-	event.Stop.AddHandler(
-		func() error {
-			srv.log.Info("Stop healthcheck listener")
-
-			err := srv.listener.Shutdown(nil)
-			if err != nil {
-				return err
-			}
-
-			srv.log.Info("Stopped healthcheck listener")
-
-			return nil
-		},
-	)
+func NewHandler() *Handler {
+	return &Handler{}
 }
 
 // Add add new HTTP healthcheck
-func Add(path string, f func() (State, string)) {
-	http.HandleFunc(path, func(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) Add(path string, f func() (State, string)) {
+	h.HandleFunc(path, func(w http.ResponseWriter, r *http.Request) {
 		state, text := f()
 
 		if state != StatePassing {
 			w.WriteHeader(stateMap[state])
 		}
 
-		_, err := fmt.Fprintf(w, text)
-		if err != nil {
-			srv.log.Error(err)
-		}
+		_, _ = fmt.Fprint(w, text)
 	})
 }
 
 // AddReq add new HTTP healthcheck with http.Request parameter
 // to allow get some data from request
-func AddReq(path string, f func(*http.Request) (State, string)) {
-	http.HandleFunc(path, func(w http.ResponseWriter, r *http.Request) {
-		srv.log.Debug("Request ", path)
-
+func (h *Handler) AddReq(path string, f func(*http.Request) (State, string)) {
+	h.HandleFunc(path, func(w http.ResponseWriter, r *http.Request) {
 		state, text := f(r)
-
-		srv.log.Debug("Response state: ", state)
 
 		if state != StatePassing {
 			w.WriteHeader(stateMap[state])
 		}
 
-		_, err := fmt.Fprintf(w, text)
-		if err != nil {
-			srv.log.Error(err)
-		}
+		_, _ = fmt.Fprint(w, text)
 	})
 }