+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2019-04-04 v0.5.3
   - Go modules.
+  - Remove request/response logging.
 
 2018-10-07 v0.5.2
   - New AddReq function to allow healthcheck get some data from request.